shared/requirements.txt
# SQLAlchemy for database configuration [cite: 13]
sqlalchemy==2.0.23

# Python JWT for shared authentication utilities 
pyjwt==2.8.0

# MinIO client for file management [cite: 10]
minio==7.2.0

# Python-dotenv for environment variable management
python-dotenv==1.0.0